nimue simulation plot

nimue simulation plot

# S3 method for class 'nimue_simulation'
plot(
  x,
  var_select = NULL,
  replicates = FALSE,
  summarise = TRUE,
  ci = TRUE,
  q = c(0.025, 0.975),
  summary_f = mean,
  x_var = "t",
  date_0 = NULL,
  particle_fit = FALSE,
  ...
)

Arguments

x

An squire_simulation object

var_select

Vector of variable names to plot (default is all)

replicates

Plot replicates

summarise

Logical, add summary line

ci

logical add confidence interval ribbon

q

Quantiles for upper and lower of interval ribbon

summary_f

Function to summarise each compartment passed to the fun argument of stat_summary

x_var

X variable to use for plotting (default is "t", but can be set to, "date", if date_0 provided), which will cause the date to be plotted rather than time.

date_0

Date of time 0 (e.g. "2020-03-01"), if specified a date column will be added

particle_fit

If the squire_simulation provided is the result of running the particle filter, do we want to just plot the fit. Default = FALSE

...

additional arguments affecting the plot produced.